Add Attendance

+++ laravel/database/migrations/2025_01_13_103959_create_employees_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('employees', function (Blueprint $table) {
             $table->id();
             $table->foreignId("user_id")->constrained("users")->onDelete("cascade");
-            $table->float("salary");
+            $table->double("salary");
             $table->string("salary_method");
             $table->date("join_date");
             $table->timestamps();
